==> 3. Matching NewPointOfConnection to effective energy offers.ipynb <==
import datetime
import numpy as np
import matplotlib.pyplot as plt
import matplotlib.dates as mdates
import matplotlib.cbook as cbook
import EAtools as ea
import pandas as pd
import os   
import sys 
import numpy
import fileinput
import matplotlib.pyplot as plt 
import matplotlib.ticker  
from matplotlib.ticker import FormatStrFormatter 
import pandas.io.sql
import pyodbc 
from matplotlib.pyplot import figure, show
from matplotlib.patches import Ellipse
import numpy as np
from pandas import *
import EAtools as ea
ea.set_options()
ea.ea_report_style()
from pylab import *
import matplotlib.ticker as tkr
from datetime import datetime, date, time, timedelta
ea_p=ea.ea_p
ea_s=ea.ea_s
def cm2inch(value):
    return value/2.54
import pyarrow as pa
import pyarrow.parquet as pq

from db import DB
import os
ea.set_options()
ea.ea_report_style()
%pylab inline
%matplotlib inline
path='/media/usb/notebooks/Julia/20192020_review'
db = DB(profile="hallj") 
# db = DB(profile="millerr") 

rogersdata = pd.read_parquet('/media/usb/notebooks/Roger/Stacked offer plots/EffectiveEnergyOffers.parquet')                      
rogersdata.head(10)

rogersdata.dtypes

rogersdata = rogersdata.rename(columns={'Trading_DATE':'TradingDate','Trading_Period':'TradingPeriod','Offer_block':'Tranche','Quantity_Calc':'Megawatt','Offer_price':'DollarsPerMegawattHour'})
rogersdata['PointOfConnection'] = rogersdata['PNode'].str.slice(0,7)
rogersdata['PointOfConnection'] = rogersdata['PointOfConnection'].str.strip()
rogersdata['Unit'] = rogersdata['PNode'].str.slice(7,13)
rogersdata['Unit'] = rogersdata['Unit'].str.strip()
#rogersdata['Offered_quantity_to_subtract'] = rogersdata['Offer_quantity'] - rogersdata['Quantity_Calc']
rogersdata = rogersdata.sort_values(['TradingDate','TradingPeriod','PointOfConnection','Unit','Tranche'])
rogersdata = rogersdata.reset_index(drop=True).set_index(['TradingDate','TradingPeriod','PointOfConnection','Unit','Tranche'])
#rogersdata = rogersdata[['Offered_quantity_to_subtract']]
rogersdata = rogersdata[['Megawatt','DollarsPerMegawattHour']]
rogersdata.head()

temp = rogersdata.reset_index()
temp.Unit.unique()

temp.PointOfConnection.unique()

rogersdata.tail()

len(rogersdata)

#match in newpointofconnection and newpointofconnection2
offers = read_csv(path+'/offers_NewPointOfConnectoin.csv')
offers['PointOfConnection'] = offers['PointOfConnection'].str.strip()
offers['Unit'] = offers['Unit'].str.strip()
offers = offers.set_index(['PointOfConnection','Unit'])
offers = offers.groupby(level=[0,1])['NewPointOfConnection','NewPointOfConnection2'].tail(1)
offers.head()

offers

rogersdata = rogersdata.reset_index().set_index(['PointOfConnection','Unit'])
rogersdata = rogersdata.join(offers,how='left')
rogersdata.head()

rogersdata = rogersdata.reset_index()
rogersdata.head()

len(rogersdata)

table = pa.Table.from_pandas(rogersdata)
pq.write_table(table, path + "/rogersdata.parquet")
