==> 4a. DOASA WaterValues dataprep.ipynb <==
import pandas as pd
import pandas.io.sql as sql
import EAtools as ea
%pylab inline
from datetime import date, datetime, time, timedelta
import pyarrow as pa
import pyarrow.parquet as pq
import matplotlib.dates as mdates
from db import DB
db = DB(profile="hallj")  # <need your username here

from IPython.html.widgets import interact, fixed
from IPython.display import clear_output, display, HTML

ea.set_options()
ea.ea_all_colours()
from matplotlib import rcParams
rcParams['font.family'] = 'monospace'
from matplotlib import pyplot as plt

import os   
import sys 
import datetime
import numpy as np

path='/media/usb/notebooks/Julia/20192020_review'

# DOASA

DOASA_watervalues = pd.read_csv('/media/P/Market Performance/MarginalH20Values/20210812_DOASA_Simulation_2016_2021_4000iter/Estimated_H20_Marginal_Values/NZ_marginal_H20_value.csv')
DOASA_watervalues['Date'] = pd.to_datetime(DOASA_watervalues.tradingdate)
DOASA_watervalues = DOASA_watervalues.reset_index(drop=True).set_index(['Date'])
DOASA_watervalues = DOASA_watervalues[['marginal_H20_value']]
DOASA_watervalues.head()

#save parquet file
watervalues_table = pa.Table.from_pandas(DOASA_watervalues)
pq.write_table(watervalues_table, "/media/usb/notebooks/Julia/20192020_review/DOASA_avg_watervalue.parquet")
